requirements.txt
# -------------------------------

# 🧠 CORE PYTHON + UTILITIES

# -------------------------------

pandas==2.3.3

numpy==2.3.4

scikit-learn==1.7.2

scipy==1.16.2

joblib==1.5.2

python-dateutil==2.9.0.post0

pytz==2025.2

tzdata==2025.2
 
# -------------------------------

# 📊 VISUALIZATION & UI

# -------------------------------

matplotlib==3.10.7

seaborn==0.13.2

plotly==6.3.1

streamlit==1.50.0
 
# -------------------------------

# 🧩 LLM & AI FRAMEWORKS

# -------------------------------

openai==1.109.1

google-genai==1.46.0

langchain==1.0.2

langchain-core==1.0.0

langchain-community==0.4

langgraph==1.0.1

langgraph-checkpoint==3.0.0

langgraph-prebuilt==1.0.1

langgraph-sdk==0.2.9

langsmith==0.4.37

crewai==1.1.0
 
# -------------------------------

# ⚙️ SYSTEM & CLOUD UTILITIES

# -------------------------------

python-dotenv==1.1.1

requests==2.32.5

protobuf==6.33.0

google-auth==2.41.1
 
# -------------------------------

# 📦 NOTEBOOK / DEV TOOLS

# -------------------------------

jupyter==1.1.1

notebook==7.4.7

ipykernel==7.0.1
 
# -------------------------------

# 🧪 TESTING

# -------------------------------

pytest==8.4.2

pytest-cov==7.0.0
 
# -------------------------------

# 🧰 OPTIONAL PERFORMANCE / OPTIMIZATION

# -------------------------------

uvicorn==0.38.0

fastapi==0.115.0

psutil==7.1.1

 